+++ core/document-service/apidoc/api_data.js
@@ -1,1 +1,1 @@
-define({ "api": [  {    "type": "delete",    "url": "/catalog/:catalog_id",    "title": "删除目录",    "group": "Catalog",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure -i -X DELETE \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "DeleteCatalogCatalog_id"  },  {    "type": "get",    "url": "/catalog/",    "title": "获取一级目录和文档(私有/公共)",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "ispublic",            "description": "<p>true/false 默认false：表示是私有目录，true：公共目录</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\t--data \"ispublic=true\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalogChilds\": [{ \n\t\t\"catalog_id\":\"d3b2ea4082724b23bbb9896506626e13\",\n\t\t\"parent_catalog_id\": null, \n\t\t\"catalog_name\": \"目录名\",\n\t\t\"isroot\": true, \n\t\t\"author\": \"admin\", \n\t\t\"opdatetime\":\"2016-05-25\" }], \n\t\"documentChilds\": [{ \n\t\t\"file_id\":\"75eec2d14e3d41b58a0e35920c4c9615\", \n\t\t\"catalog_id\": null, \n\t\t\"file_name\": \"文件名\", \n\t\t\"author\": \"admin\",\n\t\t\"opdatetime\": \"2016-05-25\" \n\t}] \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "GetCatalog"  },  {    "type": "get",    "url": "/catalog/:catalog_id",    "title": "获取目录下子目录和文件列表",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalogChilds\": [{ \n\t\t\"catalog_id\":\"d3b2ea4082724b23bbb9896506626e13\",\n\t\t\"parent_catalog_id\": null, \n\t\t\"catalog_name\": \"目录名\",\n\t\t\"isroot\": true, \n\t\t\"author\": \"admin\", \n\t\t\"opdatetime\":\"2016-05-25\" }], \n\tdocumentChilds: [{ \n\t\t\"file_id\":\"75eec2d14e3d41b58a0e35920c4c9615\", \n\t\t\"catalog_id\":null, \n\t\t\"file_name\": \"文件名\", \n\t\t\"author\": \"admin\",\n\t\t\"opdatetime\": \"2016-05-25\" \n\t}] \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "GetCatalogCatalog_id"  },  {    "type": "patch",    "url": "/catalog/:catalog_id/?catalog_name=新目录名",    "title": "修改目录名称",    "group": "Catalog",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_name",            "description": "<p>目录名称</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure -i -X DELETE \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782/?catalog_name=新目录名",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\",\n\t\t\"parent_catalog_id\": null, \n\t\t\"catalog_name\": \"目录名\",\n\t\t\"author\": \"zhanglei\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PatchCatalogCatalog_idCatalog_name"  },  {    "type": "post",    "url": "/catalog/:catalog_id/?catalog_name=目录名",    "title": "创建子目录(私有/公共)",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>上级目录ID</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_name",            "description": "<p>目录名称</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "ispublic",            "description": "<p>true/false 默认false，表示是我的目录</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\t--data \"catalog_name=目录名&ispublic=true\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\",\n\t\t\"parent_catalog_id\":\"033b04027ae2429f81f985f9cce2978c\", \n\t\t\"catalog_name\":\"子目录名\", \n\t\t\"author\": \"zhanglei\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PostCatalogCatalog_idCatalog_name"  },  {    "type": "post",    "url": "/catalog/:catalog_id/document/",    "title": "上传文档到目录",    "group": "Catalog",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -X POST -H\n\t\"Authorization: Bearer <access_token>\" \\ -F \"file=@file.doc\" \\ \n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782/document/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true,  \n\t\"document\": { \n\t\t\"file_id\":\"56bdcd1a28c948bb881cf3e9a31cd782\", \n\t\t\"catalog_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\", \n\t\t\"file_name\": \"文档.xlsx\", \n\t\t\"author\": \"zhanglei\", \n\t\t\"links\": [{ \n\t\t\t\"id\":\"view\", \n\t\t\t\"url\":\"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n     },{ \n     \t\"id\": \"info\", \n     \t\"url\":\"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n     },{ \n     \t\"id\": \"download\", \n     \t\"url\":\"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n     }] \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PostCatalogCatalog_idDocument"  },  {    "type": "post",    "url": "/catalog/?catalog_name=目录名",    "title": "创建根目录(私有/公共)",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_name",            "description": "<p>目录名称</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "ispublic",            "description": "<p>true/false 默认false，表示是我的目录</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -X POST -H\n\t\"Authorization: Bearer <access_token>\" \\\n\t--data \"catalog_name=目录名&ispublic=true\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{\n\t\"success\":true, \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\", \n\t\t\"catalog_name\":\"目录名\", \n\t\t\"author\": \"zhanglei\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PostCatalogCatalog_name"  },  {    "type": "delete",    "url": "/document/:fileid",    "title": "删除文档",    "group": "Document",    "description": "<p>删除文档会同时删除这个文档的分享</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X DELETE \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "DeleteDocumentFileid"  },  {    "type": "get",    "url": "/document/:fileid/download",    "title": "下载文档",    "group": "Document",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure \\\n\t-o file.docx \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/api/v1/documentservice/document/81bdcd1a28c948bb881cf3e9a31cd782/download",        "type": "curl"      }    ],    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentFileidDownload"  },  {    "type": "get",    "url": "/document/:fileid/info",    "title": "获取文档信息",    "group": "Document",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n    \"file_name\": \"文档.xlsx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n    }]\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentFileidInfo"  },  {    "type": "get",    "url": "/document/me",    "title": "获取我的文档列表",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/me",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"message\": null,\n  \"documents\": [{\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n    \"file_name\": \"文档.xlsx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n    }]\n  },{\n    \"file_id\": \"646a2f9a7cb34151a8cdfd618aeb3018\",\n    \"file_name\": \"文档2.docx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/download\"\n    }]\n  }]\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentMe"  },  {    "type": "get",    "url": "/document/me/count",    "title": "获取我的文档数量",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/me",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"count\": 10\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentMeCount"  },  {    "type": "get",    "url": "/document/me/download/count",    "title": "我的文档下载量",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/me/download/count",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"count\": 10\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentMeDownloadCount"  },  {    "type": "get",    "url": "/document/operate/?optype=:optype",    "title": "最近的操作",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "optype",            "description": "<p>操作类型，支持create,delete,download,info,view,share,copy,move</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-X GET \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/operate/?optype=create,download",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"documentOpLogs\": []\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentOperateOptypeOptype"  },  {    "type": "get",    "url": "/document/search",    "title": "按文件名模糊搜索",    "group": "Document",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "q",            "description": "<p>搜索条件，多个条件用加号连接，每个条件采用模糊匹配，多个条件之间采用或的关系</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/search?q=亿阳+指南",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"message\": null,\n  \"documents\": [{\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n    \"file_name\": \"亿阳企业文化.xlsx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n    }]\n  },{\n    \"file_id\": \"646a2f9a7cb34151a8cdfd618aeb3018\",\n    \"file_name\": \"开发指南.docx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/download\"\n    }]\n  }]\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentSearch"  },  {    "type": "post",    "url": "/document/",    "title": "上传文档",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\t-F \"file=@file.doc\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 201 OK\n{\n  \"success\": true,\n  \"message\": null\n  \"document\": {\n  \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n  \"file_name\": \"文档.xlsx\",\n  \"author\": \"zhanglei\",\n  \"links\": [         {\n    \"id\": \"view\",\n    \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n      }]\n   }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocument"  },  {    "type": "post",    "url": "/document/:fileid/?copyto_catalog_id=:catalog_id",    "title": "复制文档到目录",    "group": "Document",    "description": "<p>复制文档到指定目录，此时会产生一个文件副本，源文档的修改不会影响副本</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X POST \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/?copyto_catalog_id=:catalog_id",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocumentFileidCopyto_catalog_idCatalog_id"  },  {    "type": "post",    "url": "/document/:fileid/?moveto_catalog_id=:catalog_id",    "title": "移动文档到目录",    "group": "Document",    "description": "<p>将这个文档移动到指定的目录，移动后这个文档在原有目录下消失</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X POST \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/?moveto_catalog_id=:catalog_id",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocumentFileidMoveto_catalog_idCatalog_id"  },  {    "type": "post",    "url": "/document/:fileid/?shareto_catalog_id=:catalog_id",    "title": "分享文档到目录",    "group": "Document",    "description": "<p>分享这个文档的访问地址到指定目录，此文档删除后分享文档也将消失</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X POST \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/?shareto_catalog_id=:catalog_id",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocumentFileidShareto_catalog_idCatalog_id"  }] });
+define({ "api": [  {    "type": "delete",    "url": "/catalog/:catalog_id",    "title": "删除目录",    "group": "Catalog",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure -i -X DELETE \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "DeleteCatalogCatalog_id"  },  {    "type": "get",    "url": "/catalog/",    "title": "获取一级目录和文档(私有/公共)",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "ispublic",            "description": "<p>true/false 默认false：表示是私有目录，true：公共目录</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\t--data \"ispublic=true\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalogChilds\": [{ \n\t\t\"catalog_id\":\"d3b2ea4082724b23bbb9896506626e13\",\n\t\t\"parent_catalog_id\": null, \n\t\t\"catalog_name\": \"目录名\",\n\t\t\"isroot\": true, \n\t\t\"author\": \"admin\", \n\t\t\"opdatetime\":\"2016-05-25\" }], \n\t\"documentChilds\": [{ \n\t\t\"file_id\":\"75eec2d14e3d41b58a0e35920c4c9615\", \n\t\t\"catalog_id\": null, \n\t\t\"file_name\": \"文件名\", \n\t\t\"author\": \"admin\",\n\t\t\"opdatetime\": \"2016-05-25\" \n\t}] \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "GetCatalog"  },  {    "type": "get",    "url": "/catalog/:catalog_id",    "title": "获取目录下子目录和文件列表",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalogChilds\": [{ \n\t\t\"catalog_id\":\"d3b2ea4082724b23bbb9896506626e13\",\n\t\t\"parent_catalog_id\": null, \n\t\t\"catalog_name\": \"目录名\",\n\t\t\"isroot\": true, \n\t\t\"author\": \"admin\", \n\t\t\"opdatetime\":\"2016-05-25\" }], \n\tdocumentChilds: [{ \n\t\t\"file_id\":\"75eec2d14e3d41b58a0e35920c4c9615\", \n\t\t\"catalog_id\":null, \n\t\t\"file_name\": \"文件名\", \n\t\t\"author\": \"admin\",\n\t\t\"opdatetime\": \"2016-05-25\" \n\t}] \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "GetCatalogCatalog_id"  },  {    "type": "patch",    "url": "/catalog/:catalog_id/?catalog_name=新目录名",    "title": "修改目录名称",    "group": "Catalog",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_name",            "description": "<p>目录名称</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure -i -X DELETE \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782/?catalog_name=新目录名",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\",\n\t\t\"parent_catalog_id\": null, \n\t\t\"catalog_name\": \"目录名\",\n\t\t\"author\": \"zhanglei\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PatchCatalogCatalog_idCatalog_name"  },  {    "type": "post",    "url": "/catalog/:catalog_id/?catalog_name=目录名",    "title": "创建子目录(私有/公共)",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>上级目录ID</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_name",            "description": "<p>目录名称</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "ispublic",            "description": "<p>true/false 默认false，表示是我的目录</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -H\n\t\"Authorization: Bearer <access_token>\" \\\n\t--data \"catalog_name=目录名&ispublic=true\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true, \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\",\n\t\t\"parent_catalog_id\":\"033b04027ae2429f81f985f9cce2978c\", \n\t\t\"catalog_name\":\"子目录名\", \n\t\t\"author\": \"zhanglei\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PostCatalogCatalog_idCatalog_name"  },  {    "type": "post",    "url": "/catalog/:catalog_id/document/",    "title": "上传文档到目录",    "group": "Catalog",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_id",            "description": "<p>目录ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -X POST -H\n\t\"Authorization: Bearer <access_token>\" \\ -F \"file=@file.doc\" \\ \n\thttps://localhost:8000/documentservice/api/v1/catalog/81bdcd1a28c948bb881cf3e9a31cd782/document/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{ \n\t\"success\":true,  \n\t\"document\": { \n\t\t\"file_id\":\"56bdcd1a28c948bb881cf3e9a31cd782\", \n\t\t\"catalog_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\", \n\t\t\"file_name\": \"文档.xlsx\", \n\t\t\"author\": \"zhanglei\", \n\t\t\"links\": [{ \n\t\t\t\"id\":\"view\", \n\t\t\t\"url\":\"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n     },{ \n     \t\"id\": \"info\", \n     \t\"url\":\"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n     },{ \n     \t\"id\": \"download\", \n     \t\"url\":\"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n     }] \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PostCatalogCatalog_idDocument"  },  {    "type": "post",    "url": "/catalog/?catalog_name=目录名",    "title": "创建根目录(私有/公共)",    "group": "Catalog",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "catalog_name",            "description": "<p>目录名称</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "ispublic",            "description": "<p>true/false 默认false，表示是我的目录</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage: ",        "content": "curl --insecure \\ -X POST -H\n\t\"Authorization: Bearer <access_token>\" \\\n\t--data \"catalog_name=目录名&ispublic=true\" \\\n\thttps://localhost:8000/documentservice/api/v1/catalog/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response: ",          "content": "HTTP/1.1 201 OK \n{\n\t\"success\":true, \n\t\"catalog\": { \n\t\t\"catalog_id\":\"81bdcd1a28c948bb881cf3e9a31cd782\", \n\t\t\"catalog_name\":\"目录名\", \n\t\t\"author\": \"zhanglei\" \n\t} \n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response: ",          "content": "HTTP/1.1 400 Bad Request \n{\n\t\"success\": false, \n\t\"message\": \"错误信息\" \n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CatalogController.java",    "groupTitle": "Catalog",    "name": "PostCatalogCatalog_name"  },  {    "type": "get",    "url": "/document/me/collect",    "title": "我的收藏(?)",    "group": "Collect",    "description": "<p>显示我的收藏文档列表</p>",    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CollectController.java",    "groupTitle": "Collect",    "name": "GetDocumentMeCollect"  },  {    "type": "get",    "url": "/document/me/collect/count",    "title": "我的收藏量(?)",    "group": "Collect",    "description": "<p>显示我收藏的文档数量</p>",    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CollectController.java",    "groupTitle": "Collect",    "name": "GetDocumentMeCollectCount"  },  {    "type": "post",    "url": "/document/:fileid/collect",    "title": "收藏文档(?)",    "group": "Collect",    "description": "<p>收藏这个文档</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/CollectController.java",    "groupTitle": "Collect",    "name": "PostDocumentFileidCollect"  },  {    "type": "delete",    "url": "/document/:fileid",    "title": "删除文档",    "group": "Document",    "description": "<p>删除文档会同时删除这个文档的分享、并标注收藏失效</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X DELETE \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "DeleteDocumentFileid"  },  {    "type": "get",    "url": "/document/:fileid/download",    "title": "下载文档",    "group": "Document",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure \\\n\t-o file.docx \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/api/v1/documentservice/document/81bdcd1a28c948bb881cf3e9a31cd782/download",        "type": "curl"      }    ],    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentFileidDownload"  },  {    "type": "get",    "url": "/document/:fileid/info",    "title": "获取文档信息",    "group": "Document",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n    \"file_name\": \"文档.xlsx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n    }]\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentFileidInfo"  },  {    "type": "get",    "url": "/document/operate/?optype=:optype",    "title": "最近的操作",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "optype",            "description": "<p>操作类型，支持create,delete,download,info,view,share,copy,move</p>"          }        ]      }    },    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-X GET \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/operate/?optype=create,download",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"documentOpLogs\": []\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "GetDocumentOperateOptypeOptype"  },  {    "type": "post",    "url": "/document/",    "title": "上传文档",    "group": "Document",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\t-F \"file=@file.doc\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 201 OK\n{\n  \"success\": true,\n  \"message\": null\n  \"document\": {\n  \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n  \"file_name\": \"文档.xlsx\",\n  \"author\": \"zhanglei\",\n  \"links\": [         {\n    \"id\": \"view\",\n    \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n      }]\n   }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocument"  },  {    "type": "post",    "url": "/document/:fileid/?copyto_catalog_id=:catalog_id",    "title": "复制文档到目录",    "group": "Document",    "description": "<p>复制文档到指定目录，此时会产生一个文件副本，源文档的修改不会影响副本</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X POST \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/?copyto_catalog_id=:catalog_id",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocumentFileidCopyto_catalog_idCatalog_id"  },  {    "type": "post",    "url": "/document/:fileid/?moveto_catalog_id=:catalog_id",    "title": "移动文档到目录",    "group": "Document",    "description": "<p>将这个文档移动到指定的目录，移动后这个文档在原有目录下消失</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X POST \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/?moveto_catalog_id=:catalog_id",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocumentFileidMoveto_catalog_idCatalog_id"  },  {    "type": "post",    "url": "/document/:fileid/?shareto_catalog_id=:catalog_id",    "title": "分享文档到目录",    "group": "Document",    "description": "<p>分享这个文档的访问地址到指定目录，此文档删除后分享文档也将消失</p>",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "fileid",            "description": "<p>文档ID</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i -X POST \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/?shareto_catalog_id=:catalog_id",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 204 NO CONTENT\n{\n  \"success\": true,\n  \"message\": null,\n  \"document\": {\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\"\n  }\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 404 Not Found\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/DocumentController.java",    "groupTitle": "Document",    "name": "PostDocumentFileidShareto_catalog_idCatalog_id"  },  {    "type": "get",    "url": "/document/search",    "title": "按文件名模糊搜索",    "group": "Search",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "q",            "description": "<p>搜索条件，多个条件用加号连接，每个条件采用模糊匹配，多个条件之间采用或的关系</p>"          }        ]      }    },    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/search?q=亿阳+指南",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"message\": null,\n  \"documents\": [{\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n    \"file_name\": \"亿阳企业文化.xlsx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n    }]\n  },{\n    \"file_id\": \"646a2f9a7cb34151a8cdfd618aeb3018\",\n    \"file_name\": \"开发指南.docx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/download\"\n    }]\n  }]\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/SearchController.java",    "groupTitle": "Search",    "name": "GetDocumentSearch"  },  {    "type": "get",    "url": "/document/me",    "title": "获取我的文档列表",    "group": "Statistic",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/me",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"message\": null,\n  \"documents\": [{\n    \"file_id\": \"81bdcd1a28c948bb881cf3e9a31cd782\",\n    \"file_name\": \"文档.xlsx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/81bdcd1a28c948bb881cf3e9a31cd782/download\"\n    }]\n  },{\n    \"file_id\": \"646a2f9a7cb34151a8cdfd618aeb3018\",\n    \"file_name\": \"文档2.docx\",\n    \"author\": \"zhanglei\",\n    \"links\": [{\n      \"id\": \"view\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/view\"\n    },{\n      \"id\": \"info\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/info\"\n    },{\n      \"id\": \"download\",\n      \"url\": \"/documentservice/api/v1/document/646a2f9a7cb34151a8cdfd618aeb3018/download\"\n    }]\n  }]\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/StatisticController.java",    "groupTitle": "Statistic",    "name": "GetDocumentMe"  },  {    "type": "get",    "url": "/document/me/count",    "title": "获取我的文档数量",    "group": "Statistic",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/me",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"count\": 10\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/StatisticController.java",    "groupTitle": "Statistic",    "name": "GetDocumentMeCount"  },  {    "type": "get",    "url": "/document/me/download/count",    "title": "我的文档下载量",    "group": "Statistic",    "permission": [      {        "name": "none"      }    ],    "examples": [      {        "title": "Example usage:",        "content": "curl --insecure -i \\\n\t-H \"Authorization: Bearer <access_token>\" \\\n\thttps://localhost:8000/documentservice/api/v1/document/me/download/count",        "type": "curl"      }    ],    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": true,\n  \"count\": 10\n}",          "type": "json"        }      ]    },    "error": {      "examples": [        {          "title": "Error-Response:",          "content": "HTTP/1.1 400 Bad Request\n{\n  \"success\": false,\n  \"message\": \"错误信息\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/StatisticController.java",    "groupTitle": "Statistic",    "name": "GetDocumentMeDownloadCount"  },  {    "type": "post",    "url": "/document/star",    "title": "热门资源(?)",    "group": "Statistic",    "description": "<p>文档评级最高的资源</p>",    "version": "0.0.0",    "filename": "src/main/java/org/iplatform/microservices/core/documentservice/controller/StatisticController.java",    "groupTitle": "Statistic",    "name": "PostDocumentStar"  }] });
