+++ core/document-service/apidoc/api_project.json
@@ -1,1 +1,1 @@
-{  "name": "文档服务 REST API",  "version": "0.1.0",  "description": "文档服务",  "title": "文档服务 RESTAPI",  "url": "https://localhost:8000/documentservice/api/v1",  "sampleUrl": false,  "apidoc": "0.2.0",  "generator": {    "name": "apidoc",    "time": "2016-05-26T13:25:04.364Z",    "url": "http://apidocjs.com",    "version": "0.16.1"  }}
+{  "name": "文档服务 REST API",  "version": "0.1.0",  "description": "文档服务",  "title": "文档服务 RESTAPI",  "url": "https://localhost:8000/documentservice/api/v1",  "sampleUrl": false,  "apidoc": "0.2.0",  "generator": {    "name": "apidoc",    "time": "2016-05-30T02:51:15.265Z",    "url": "http://apidocjs.com",    "version": "0.16.1"  }}
